Add comments examples to all commentable resources
