+++ decidim-budgets/spec/features/comments_spec.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+require "spec_helper"
+
+describe "Comments", type: :feature do
+  let!(:feature) { create(:budget_feature, organization: organization) }
+  let!(:commentable) { create(:project, feature: feature) }
+
+  let(:resource_path) { decidim_budgets.project_path(commentable, feature_id: feature, participatory_process_id: feature.participatory_process) }
+  include_examples "comments"
+end
