+++ decidim-comments/app/commands/decidim/comments/create_comment.rb
@@ -22,7 +22,7 @@ module Decidim
         return broadcast(:invalid) if form.invalid?
 
         create_comment
-        send_notification_to_author unless @author == @commentable.author
+        send_notification_to_author if has_author? && !same_author?
 
         broadcast(:ok, @comment)
       end
@@ -45,6 +45,14 @@ module Decidim
         elsif @comment.depth == 0 && @commentable.author.comments_notifications?
           CommentNotificationMailer.comment_created(@author, @comment, @commentable).deliver_later
         end
+      end
+
+      def has_author?
+        @commentable.respond_to?(:author) && @commentable.author.present?
+      end
+
+      def same_author?
+        @author == @commentable.author
       end
     end
   end
