+++ decidim-core/lib/decidim/core/test.rb
@@ -6,3 +6,4 @@ require "decidim/core/test/shared_examples/has_feature"
 require "decidim/core/test/shared_examples/has_scope"
 require "decidim/core/test/shared_examples/has_category"
 require "decidim/core/test/shared_examples/has_reference"
+require "decidim/core/test/shared_examples/comments_examples"
